Fixed character crash during gameplay
Created an cpp file named "Functions" to hold all the functions while the existing header file holds the function prototypes

Fixed bug during gameplay where the game crashes if characters are entered instead of numbers

Created an input validation function to call instead of rewriting the same code over and over again

created an end screed function to call instead of rewriting the same code over and over again

added parameters to multiplayer and single player functions to avoid excessive code repetitions
