+++ Tic-Tac-Toe Game/Final_Game_With_Bug_Fixes/tic tac toe/tic tac toe/Debug/tic tac toe.log
@@ -1,2 +1,1 @@
-﻿  Source.cpp
-  tic tac toe.vcxproj -> C:\Users\cnken\Desktop\Software-Design\Tic-Tac-Toe Game\Final_Game_With_Bug_Fixes\tic tac toe\Debug\tic tac toe.exe
+﻿C:\Program Files (x86)\Microsoft Visual Studio\2017\Community\Common7\IDE\VC\VCTargets\Microsoft.Cpp.Platform.targets(57,5): error MSB8020: The build tools for v142 (Platform Toolset = 'v142') cannot be found. To build using the v142 build tools, please install v142 build tools.  Alternatively, you may upgrade to the current Visual Studio tools by selecting the Project menu or right-click the solution, and then selecting "Retarget solution".
