+++ Tic-Tac-Toe Game/Final_Game_With_Bug_Fixes/tic tac toe/tic tac toe/Header.h
@@ -5,625 +5,14 @@
 #include <string>
 using namespace std;
 
-//FUNCTION PROTOTYPES
-void singlePlayer();
-void multiPlayer();
-void position(string [][3], int, string, int);
-bool checkWinner(string [][3], string);
-void comPosition(string [][3], int, string);
-
-
-//FUNCTION TO DISPLAY THE GAME BOARD
-void gameBoard(string board[][3]) {
-
-	const int ROWS = 3, COLS = 3;		// Set the number of rows and columns
-
-										//TIC TAC TOE BOARD
-	for (int r = 0; r < ROWS; r++) {							//For loop to create the rows of the board (3 rows)
-
-		for (int c = 0; c < COLS; c++) {						//For loop to create the columns for each row (3 columns per row)
-			cout << setw(3) << " " << board[r][c] << " ";		//Out put the numbers from the board array onto the location specified
-
-			if (c != 2) {
-				cout << "|";									//Divide the columns with a "|"
-			}
-		}
-
-		if (r != 2) {
-			cout << endl << setw(3) << "_________________________" << endl;			//Divide the rows with an underscore
-		}
-	}
-
-
-}
-
-void displayMenu() {		//display the main menu
-	int menuChoice;						//int to hold the player's input choice in the main menu
-
-										//MAIN MENU
-	cout << "MAIN MENU" << endl;
-	cout << "Tic Tac Toe" << endl;
-	cout << "Rules:" << endl;
-	cout << "-Player 1 is X and Player 2 is O" << endl;
-	cout << "-The game will prompt which player's turn it is. The player will choose an available number from the board" << endl;
-	cout << "-The game ends when a player gets three in a row or when the board is completely filled" << endl;
-	cout << "-The player who gets three in a row wins" << endl << endl;
-
-	cout << "1 - Single Player" << endl;		//These are the player's options
-	cout << "2 - Multiplayer" << endl;			//
-	cout << "3 - Quit" << endl << endl;			//
-
-	cout << "(Enter 1 to play single player, 2 for multiplayer, or 3 to quit the program)" << endl;
-	cin >> menuChoice;
-
-	//Input Validation
-	while (menuChoice < 1 || menuChoice > 3 || cin.fail()) {			//while loop to make sure the player inputs a valid digit
-		if (cin.fail()) {		//If the input is not a number, clear the error
-			cin.clear();
-			cin.ignore(numeric_limits<streamsize>::max(), '\n');
-		}
-		cout << "This is an invalid input!" << endl;
-		cout << "Enter 1 to play single player, 2 for multiplayer, or 3 to quit the program" << endl;
-		cin >> menuChoice;
-	}
-
-	if (menuChoice == 1) {
-		singlePlayer();
-	}
-	else if (menuChoice == 2) {
-		multiPlayer();
-	}
-	else {
-		exit(0);		//if player input is 3, close the program
-	}
-
-	cout << endl;
-}
-
-void singlePlayer() {		//begin single player mode
-	cout << "SINGLE PLAYER" << endl;
-
-	bool p1Turn;								//Boolean variables used to switch between player turns								
-	bool comTurn;								//
-
-	string player[2] = { "X", "O" };			//array that holds player letter
-
-	bool gameOver = false;	//determines if the game is over
-
-	int counter = 9;		//counter to keep track of how many spaces left (9 at the beginning)
-
-	int p1Choice, comChoice;			//variables to hold the player's choice and the computer's choice
-
-	const int ROWS = 3, COLS = 3;		// Set the number of rows and columns
-	string board[ROWS][COLS] = { "(1)", "(2)", "(3)",			//2 dimensional array to create tic tac toe board
-		"(4)", "(5)", "(6)",
-		"(7)", "(8)", "(9)" };
-
-	p1Turn = true;				//Set p1turn to true to begin the game
-
-	do {
-		if (p1Turn == true) {
-			gameBoard(board);			//Call gameBoard function to display tic tac toe board
-			cout << endl;
-
-			cout << "Player " << player[0] << " choose an available spot on the board" << endl;
-			cin >> p1Choice;			//Player 1 inputs desired spot on board
-
-			while (p1Choice < 1 || p1Choice > 9) {		//Input validation (make sure player chooses a spot on the board
-				cout << "This is not a valid spot. Please input a number from 1-9.";
-				cin >> p1Choice;
-			}
-
-			position(board, p1Choice, player[0], counter);		//Call the position function to place the player on the desired position on the board
-
-			gameOver = checkWinner(board, player[0]);	//Call the checkWinner function to see if there is a winner
-
-			if (counter == 0) {
-				gameOver = true;
-				cout << "It is a tie" << endl;
-			}
-
-			if (!gameOver) {							//Check if the game is not over. If it is not, then the game continues
-				p1Turn = false;
-				comTurn = true;
-			}
-		}
-		else if (comTurn == true && counter != 0) {
-			cout << endl;
-
-			comChoice = rand() % 9 + 1;			//computer chooses what number it desires
-
-			comPosition(board, comChoice, player[1]);		//call the comPosition function to place the computer on the desired position
-
-			gameOver = checkWinner(board, player[1]);	//Call the checkWinner function to see if there is a winner
-
-			if (!gameOver) {
-				p1Turn = true;
-				comTurn = false;
-			}
-		}
-	} while (gameOver == false);
-
-	cout << endl << endl;
-	int c;		//int to hold player's choice
-
-	cout << "Enter 1 to go to the main menu or 2 to quit the program" << endl;
-	cin >> c;
-
-	//INPUT VALIDATION
-	while (c < 1 || c > 2 || cin.fail()) {
-		if (cin.fail()) {	//if input is not a number, clear the error
-			cin.clear();
-			cin.ignore(numeric_limits<streamsize>::max(), '\n');
-		}
-		cout << "This is an invalid input!" << endl;
-		cout << "Enter 1 to go to the main menu or 2 to quit the program" << endl;
-		cin >> c;
-
-	}
-
-	if (c == 1) {
-		displayMenu();	//display the menu if the player input is 1
-	}
-	else {
-		exit(0);		//close the program if player input is 2
-	}
-}
-
-void multiPlayer() {		//begin multiplayer mode
-	int p1Choice, p2Choice;
-	bool p1Turn;								//Boolean variables used to switch between player turns
-	bool p2Turn;								//
-
-	string player[2] = { "X", "O" };			//array that holds player letter
-
-	bool gameOver = false;	//determines if the game is over
-
-	cout << "MULTIPLAYER" << endl;
-	int counter = 9;		//counter to keep track of how many spaces left (9 at the beginning)
-
-	const int ROWS = 3, COLS = 3;		// Set the number of rows and columns
-	string board[ROWS][COLS] = { "(1)", "(2)", "(3)",			//2 dimensional array to create tic tac toe board
-		"(4)", "(5)", "(6)",
-		"(7)", "(8)", "(9)" };
-
-	p1Turn = true;				//Set p1turn to true to begin the game
-
-	do {			//while loop to keep the game in session
-
-		if (p1Turn == true) {		//if/else statement to switch between players during the game
-
-			gameBoard(board);			//Call gameBoard function to display tic tac toe board
-			cout << endl;
-
-			cout << "Player " << player[0] << " choose an available spot on the board" << endl;
-			cin >> p1Choice;			//Player 1 inputs desired spot on board
-
-			while (p1Choice < 1 || p1Choice > 9) {		//Input validation (make sure player chooses a spot on the board
-				cout << "This is not a valid spot. Please input a number from 1-9.";
-				cin >> p1Choice;
-			}
-
-			position(board, p1Choice, player[0], counter);		//Call the position function to place the player on the desired position on the board
-
-			gameOver = checkWinner(board, player[0]);	//Call the checkWinner function to see if there is a winner
-
-			if (counter == 0) {
-				gameOver = true;
-				cout << "It is a tie" << endl;
-			}
-
-			if (!gameOver) {							//Check if the game is not over. If it is not, then the game continues
-				p1Turn = false;
-				p2Turn = true;
-			}
-
-		}
-		else if (p2Turn == true) {
-
-			gameBoard(board);			//Call gameBoard function to display tic tac toe board
-			cout << endl;
-
-			cout << "Player " << player[1] << " choose an available spot on the board" << endl;
-			cin >> p2Choice;			//Player 1 inputs desired spot on board
-
-			while (p2Choice < 1 || p2Choice > 9) {		//Input validation (make sure player chooses a spot on the board
-				cout << "This is not a valid spot. Please input a number from 1-9.";
-				cin >> p2Choice;
-			}
-
-			position(board, p2Choice, player[1], counter);		//Call the position function to place the player on the desired position on the board
-
-			gameOver = checkWinner(board, player[1]);			//Call the checkWinner function to see if there is a winner
-
-			if (counter == 0) {
-				gameOver = true;
-				cout << "It is a tie" << endl;
-			}
-
-			if (!gameOver) {
-				p1Turn = true;
-				p2Turn = false;
-			}
-
-		}
-	} while (gameOver == false);
-
-	cout << endl << endl;
-	int c;		//int to hold player's choice
-
-	cout << "Enter 1 to go to the main menu or 2 to quit the program" << endl;
-	cin >> c;
-
-	//INPUT VALIDATION
-	while (c < 1 || c > 2 || cin.fail()) {
-		if (cin.fail()) {	//if input is not a number, clear the error
-			cin.clear();
-			cin.ignore(numeric_limits<streamsize>::max(), '\n');
-		}
-		cout << "This is an invalid input!" << endl;
-		cout << "Enter 1 to go to the main menu or 2 to quit the program" << endl;
-		cin >> c;
-
-	}
-
-	if (c == 1) {
-		displayMenu();	//display the menu if the player input is 1
-	}
-	else {
-		exit(0);		//close the program if player input is 2
-	}
-}
-
-void position(string board[][3], int pos, string player, int counter) {
-
-
-	switch (pos) {						//switch statement to determine which tile on the board to fill in depending on pos parameter
-	case 1:
-		if (board[0][0] == "[X]" || board[0][0] == "[O]") {				//if statement to make sure the spot isn't already filled
-																		//Spot is filled
-			cout << "This spot is already filled. Please choose an available spot" << endl;
-			cin >> pos;
-
-			position(board, pos, player, counter);		//Call the position function again with the new choice
-		}
-		else {
-			//Spot is available
-			if (player == "X") {			//If player X chose the spot, fill in 'X'
-				board[0][0] = "[X]";
-				counter--;					//subtract one from counter because a place on the board has been filled
-			}
-			else if (player == "O") {		//If player O chose the spot, fill in 'O'
-				board[0][0] = "[O]";
-				counter--;					//subtract one from counter because a place on the board has been filled
-			}
-		}
-		break;
-	case 2:
-		if (board[0][1] == "[X]" || board[0][1] == "[O]") {				//if statement to make sure the spot isn't already filled
-																		//Spot is filled
-			cout << "This spot is already filled. Please choose an available spot" << endl;
-			cin >> pos;
-
-			position(board, pos, player, counter);		//Call the position function again with the new choice
-		}
-		else {
-			//Spot is available
-			if (player == "X") {			//If player X chose the spot, fill in 'X'
-				board[0][1] = "[X]";
-				counter--;					//subtract one from counter because a place on the board has been filled
-			}
-			else if (player == "O") {		//If player O chose the spot, fill in 'O'
-				board[0][1] = "[O]";
-				counter--;					//subtract one from counter because a place on the board has been filled
-			}
-		}
-		break;
-	case 3:
-		if (board[0][2] == "[X]" || board[0][2] == "[O]") {				//if statement to make sure the spot isn't already filled
-																		//Spot is filled
-			cout << "This spot is already filled. Please choose an available spot" << endl;
-			cin >> pos;
-
-			position(board, pos, player, counter);		//Call the position function again with the new choice
-		}
-		else {
-			//Spot is available
-			if (player == "X") {			//If player X chose the spot, fill in 'X'
-				board[0][2] = "[X]";
-				counter--;					//subtract one from counter because a place on the board has been filled
-			}
-			else if (player == "O") {		//If player O chose the spot, fill in 'O'
-				board[0][2] = "[O]";
-				counter--;					//subtract one from counter because a place on the board has been filled
-			}
-		}
-		break;
-	case 4:
-		if (board[1][0] == "[X]" || board[1][0] == "[O]") {				//if statement to make sure the spot isn't already filled
-																		//Spot is filled
-			cout << "This spot is already filled. Please choose an available spot" << endl;
-			cin >> pos;
-
-			position(board, pos, player, counter);		//Call the position function again with the new choice
-		}
-		else {
-			//Spot is available
-			if (player == "X") {			//If player X chose the spot, fill in 'X'
-				board[1][0] = "[X]";
-				counter--;					//subtract one from counter because a place on the board has been filled
-			}
-			else if (player == "O") {		//If player O chose the spot, fill in 'O'
-				board[1][0] = "[O]";
-				counter--;					//subtract one from counter because a place on the board has been filled
-			}
-		}
-		break;
-	case 5:
-		if (board[1][1] == "[X]" || board[1][1] == "[O]") {				//if statement to make sure the spot isn't already filled
-																		//Spot is filled
-			cout << "This spot is already filled. Please choose an available spot" << endl;
-			cin >> pos;
-
-			position(board, pos, player, counter);		//Call the position function again with the new choice
-		}
-		else {
-			//Spot is available
-			if (player == "X") {			//If player X chose the spot, fill in 'X'
-				board[1][1] = "[X]";
-				counter--;					//subtract one from counter because a place on the board has been filled
-			}
-			else if (player == "O") {		//If player O chose the spot, fill in 'O'
-				board[1][1] = "[O]";
-				counter--;					//subtract one from counter because a place on the board has been filled
-			}
-		}
-		break;
-	case 6:
-		if (board[1][2] == "[X]" || board[1][2] == "[O]") {				//if statement to make sure the spot isn't already filled
-																		//Spot is filled
-			cout << "This spot is already filled. Please choose an available spot" << endl;
-			cin >> pos;
-
-			position(board, pos, player, counter);		//Call the position function again with the new choice
-		}
-		else {
-			//Spot is available
-			if (player == "X") {			//If player X chose the spot, fill in 'X'
-				board[1][2] = "[X]";
-				counter--;					//subtract one from counter because a place on the board has been filled
-			}
-			else if (player == "O") {		//If player O chose the spot, fill in 'O'
-				board[1][2] = "[O]";
-				counter--;					//subtract one from counter because a place on the board has been filled
-			}
-		}
-		break;
-	case 7:
-		if (board[2][0] == "[X]" || board[2][0] == "[O]") {				//if statement to make sure the spot isn't already filled
-																		//Spot is filled
-			cout << "This spot is already filled. Please choose an available spot" << endl;
-			cin >> pos;
-
-			position(board, pos, player, counter);		//Call the position function again with the new choice
-		}
-		else {
-			//Spot is available
-			if (player == "X") {			//If player X chose the spot, fill in 'X'
-				board[2][0] = "[X]";
-				counter--;					//subtract one from counter because a place on the board has been filled
-			}
-			else if (player == "O") {		//If player O chose the spot, fill in 'O'
-				board[2][0] = "[O]";
-				counter--;					//subtract one from counter because a place on the board has been filled
-			}
-		}
-		break;
-	case 8:
-		if (board[2][1] == "[X]" || board[2][1] == "[O]") {				//if statement to make sure the spot isn't already filled
-																		//Spot is filled
-			cout << "This spot is already filled. Please choose an available spot" << endl;
-			cin >> pos;
-
-			position(board, pos, player, counter);		//Call the position function again with the new choice
-		}
-		else {
-			//Spot is available
-			if (player == "X") {			//If player X chose the spot, fill in 'X'
-				board[2][1] = "[X]";
-				counter--;					//subtract one from counter because a place on the board has been filled
-			}
-			else if (player == "O") {		//If player O chose the spot, fill in 'O'
-				board[2][1] = "[O]";
-				counter--;					//subtract one from counter because a place on the board has been filled
-			}
-		}
-		break;
-	case 9:
-		if (board[2][2] == "[X]" || board[2][2] == "[O]") {				//if statement to make sure the spot isn't already filled
-																		//Spot is filled
-			cout << "This spot is already filled. Please choose an available spot" << endl;
-			cin >> pos;
-
-			position(board, pos, player, counter);		//Call the position function again with the new choice
-		}
-		else {
-			//Spot is available
-			if (player == "X") {			//If player X chose the spot, fill in 'X'
-				board[2][2] = "[X]";
-				counter--;					//subtract one from counter because a place on the board has been filled
-			}
-			else if (player == "O") {		//If player O chose the spot, fill in 'O'
-				board[2][2] = "[O]";
-				counter--;					//subtract one from counter because a place on the board has been filled
-			}
-		}
-		break;
-	}
-
-}
-
-void comPosition(string board[][3], int pos, string player) {
-	switch (pos) {						//switch statement to determine which tile on the board to fill in depending on pos parameter
-	case 1:
-		if (board[0][0] == "[X]" || board[0][0] == "[O]") {				//if statement to make sure the spot isn't already filled
-																		//Spot is filled
-			pos = rand() % 9 + 1;	//Make computer choose a new number
-
-			comPosition(board, pos, player);		//Call the position function again with the new choice
-		}
-		else {
-			//Spot is available
-			board[0][0] = "[O]";
-			cout << "The computer chose spot " << pos << endl;
-		}
-		break;
-	case 2:
-		if (board[0][1] == "[X]" || board[0][1] == "[O]") {				//if statement to make sure the spot isn't already filled
-																		//Spot is filled
-			pos = rand() % 9 + 1;	//Make computer choose a new number
-
-			comPosition(board, pos, player);		//Call the position function again with the new choice
-		}
-		else {
-			//Spot is available
-			board[0][1] = "[O]";
-			cout << "The computer chose spot " << pos << endl;
-		}
-		break;
-	case 3:
-		if (board[0][2] == "[X]" || board[0][2] == "[O]") {				//if statement to make sure the spot isn't already filled
-																		//Spot is filled
-			pos = rand() % 9 + 1;	//Make computer choose a new number
-
-			comPosition(board, pos, player);		//Call the position function again with the new choice
-		}
-		else {
-			//Spot is available
-			board[0][2] = "[O]";
-			cout << "The computer chose spot " << pos << endl;
-		}
-		break;
-	case 4:
-		if (board[1][0] == "[X]" || board[1][0] == "[O]") {				//if statement to make sure the spot isn't already filled
-																		//Spot is filled
-			pos = rand() % 9 + 1;	//Make computer choose a new number
-
-			comPosition(board, pos, player);		//Call the position function again with the new choice
-		}
-		else {
-			//Spot is available
-			board[1][0] = "[O]";
-			cout << "The computer chose spot " << pos << endl;
-		}
-		break;
-	case 5:
-		if (board[1][1] == "[X]" || board[1][1] == "[O]") {				//if statement to make sure the spot isn't already filled
-																		//Spot is filled
-			pos = rand() % 9 + 1;	//Make computer choose a new number
-
-			comPosition(board, pos, player);		//Call the position function again with the new choice
-		}
-		else {
-			//Spot is available
-			board[1][1] = "[O]";
-			cout << "The computer chose spot " << pos << endl;
-		}
-		break;
-	case 6:
-		if (board[1][2] == "[X]" || board[1][2] == "[O]") {				//if statement to make sure the spot isn't already filled
-																		//Spot is filled
-			pos = rand() % 9 + 1;	//Make computer choose a new number
-
-			comPosition(board, pos, player);		//Call the position function again with the new choice
-		}
-		else {
-			//Spot is available
-			board[1][2] = "[O]";
-			cout << "The computer chose spot " << pos << endl;
-		}
-		break;
-	case 7:
-		if (board[2][0] == "[X]" || board[2][0] == "[O]") {				//if statement to make sure the spot isn't already filled
-																		//Spot is filled
-			pos = rand() % 9 + 1;	//Make computer choose a new number
-
-			comPosition(board, pos, player);		//Call the position function again with the new choice
-		}
-		else {
-			//Spot is available
-			board[2][0] = "[O]";
-			cout << "The computer chose spot " << pos << endl;
-		}
-		break;
-	case 8:
-		if (board[2][1] == "[X]" || board[2][1] == "[O]") {				//if statement to make sure the spot isn't already filled
-																		//Spot is filled
-			pos = rand() % 9 + 1;	//Make computer choose a new number
-
-			comPosition(board, pos, player);		//Call the position function again with the new choice
-		}
-		else {
-			//Spot is available
-			board[2][1] = "[O]";
-			cout << "The computer chose spot " << pos << endl;
-		}
-		break;
-	case 9:
-		if (board[2][2] == "[X]" || board[2][2] == "[O]") {				//if statement to make sure the spot isn't already filled
-																		//Spot is filled
-			pos = rand() % 9 + 1;	//Make computer choose a new number
-
-			comPosition(board, pos, player);		//Call the position function again with the new choice
-		}
-		else {
-			//Spot is available
-			board[2][2] = "[O]";
-			cout << "The computer chose spot " << pos << endl;
-		}
-		break;
-	}
-}
-
-bool checkWinner(string board[][3], string player) {
-
-	string p = "[" + player + "]";		//variable used so that the if statements can recognize the filled positions
-
-	if (board[0][0] == p && board[0][1] == p && board[0][2] == p) {		//if/else statements to check if the winning combinations are met by current player
-		cout << "Player " << player << " has won!" << endl;
-		return true;					//return true and assign to gameOver variable
-	}
-	else if (board[0][0] == p && board[1][0] == p && board[2][0] == p) {
-		cout << "Player " << player << " has won!" << endl;
-		return true;
-	}
-	else if (board[0][0] == p && board[1][1] == p && board[2][2] == p) {
-		cout << "Player " << player << " has won!" << endl;
-		return true;
-	}
-	else if (board[0][1] == p && board[1][1] == p && board[2][1] == p) {
-		cout << "Player " << player << " has won!" << endl;
-		return true;
-	}
-	else if (board[0][2] == p && board[1][1] == p && board[2][0] == p) {
-		cout << "Player " << player << " has won!" << endl;
-		return true;
-	}
-	else if (board[0][2] == p && board[1][2] == p && board[2][2] == p) {
-		cout << "Player " << player << " has won!" << endl;
-		return true;
-	}
-	else if (board[1][0] == p && board[1][1] == p && board[1][2] == p) {
-		cout << "Player " << player << " has won!" << endl;
-		return true;
-	}
-	else if (board[2][0] == p && board[2][1] == p && board[2][2] == p) {
-		cout << "Player " << player << " has won!" << endl;
-		return true;
-	}
-	else {		//If there is no winner, switch players (next player's turn)
-		return false;		//return false and assign to gameOver variable
-
-	}
-}
+//function prototypes
+int inputValidate(int, int, int);
+void singlePlayer(string[], bool, bool);
+void multiPlayer(string[], bool, bool);
+void displayMenu();
+void gameBoard(string[][3]);
+void position(string[][3], int, string, int);
+bool checkWinner(string[][3], string);
+void comPosition(string[][3], int, string);
+void endScreen();
