+++ Tic-Tac-Toe Game/Final_Game_With_Bug_Fixes/tic tac toe/tic tac toe/tic tac toe.vcxproj
@@ -115,6 +115,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="Functions.cpp" />
     <ClCompile Include="Source.cpp" />
   </ItemGroup>
   <ItemGroup>
