+++ Tic-Tac-Toe Game/Final_Game_With_Bug_Fixes/tic tac toe/tic tac toe/tic tac toe.vcxproj.filters
@@ -18,6 +18,9 @@
     <ClCompile Include="Source.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="Functions.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Header.h">
